experiment: store StandardStream inside Logger, under a mutex
I thought this might be faster than re-creating a StandardStream object
for every log message, but it's not.

Implementers of Log must be Send+Sync, which RefCell isn't, so I had to
use Mutex instead to give interior mutability to the StandardStream.

It works, but isn't any faster than the old way, so there's no use for
the added complexity.