breaking: make the builder functions take self by value rather than &mut
Returning &mut Self from builder functions kind of sucks, because it
means you can't create logger and call builder functions on it in the
same line, e.g.

error[E0507]: cannot move out of a mutable reference
  --> examples/demo.rs:43:5
   |
43 |     Logger::with_verbosity(level).full_filename(false).init();
   |     ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ move occurs because
         value has type `yall::Logger`, which does not implement the `Copy` trait

It'd be nice if rustc would allow this since there's really no moves
required in the line above, but alas it complains anyway. I guess that's
because the borrow checker can't know that the &mut Logger returned is
the same as the &mut self reference passed in.

Make this nicer to use by taking in and returning self by value rather
than reference, so that users can do

    Logger::with_level(...).color(...).full_filename(...).init();

rather than

    let mut logger = Logger::with_level(...);
    logger.color(...).full_filename(...);
    logger.init();