use a Mutex for the output stream for better multi-threading support
When multiple threads are logging at the same time, output can get
pretty jumbled. Use a mutex so that only one thread can be printing
a log message at a time, keeping things readable.

Since std::sync::Mutex contains a value, we might as well put the
termcolor StandardStream in there. Level checks and message
pre-processing is all done outside of the critical region so that the
lock is held for only the minimum time.

This is like 1% slower thatn without the Mutex, which isn't a big enough
deal to be worth moving it under a feature flag.